# frontend/lib/api/openapi.json
{
  "components": {
    "schemas": {
...
      "RunDetailResponse": {
        "description": "The full body of `GET /runs/{id}` — every list-item field, plus the two that are too\nlarge or too rarely needed for a list view.\n\nA subclass of `RunListItemResponse`, the same relationship `WebsiteListItemResponse`\nhas to `WebsiteResponse`: the detail view is the list item plus more, not an\nindependently maintained sibling that could drift from it field by field.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "llms_txt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llms Txt"
          },
          "started_at": {
            "format": "date-time",
            "title": "Started At",
            "type": "string"
          },
          "stats": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stats"
          },
          "status": {
            "enum": [
              "pending",
              "processing",
              "completed",
              "failed"
            ],
            "title": "Status",
            "type": "string"
          },
          "storage_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storage Path"
          },
          "trigger": {
            "enum": [
              "manual",
              "scheduled"
            ],
            "title": "Trigger",
            "type": "string"
          },
          "website_id": {
            "format": "uuid",
            "title": "Website Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "website_id",
          "status",
          "trigger",
          "started_at",
          "completed_at",
          "duration_ms",
          "stats",
          "error",
          "llms_txt",
          "storage_path"
        ],
        "title": "RunDetailResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Supabase access token, verified locally against the project's JWKS.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for llms-text: website registration, crawl runs, llms.txt generation",
    "title": "llms-text API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs/{id}": {
      "get": {
        "description": "Return one run in full, including `llms_txt` and `storage_path`.\n\nAny signed-in user may read any run (ARCHITECTURE.md §4.1); `user_id` is present only\nto require a token and is deliberately never passed to the service. `404` if `id` names\nno run.",
        "operationId": "get_run_runs__id__get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{id}/llms-full.txt": {
      "get": {
        "description": "Download one run's `llms-full.txt` expansion as `text/plain`. Identical contract to\n`get_run_llms_txt` above, including both `404`s — see `RunService.get_llms_full_txt`.",
        "operationId": "get_run_llms_full_txt_runs__id__llms_full_txt_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Either there is no run with that id, or that run has not produced this artifact — it is still pending or processing, or it failed. The two cases answer with different `detail` strings; the status is the same because from a client's side both mean there is nothing at this URL."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Llms Full Txt",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{id}/llms.txt": {
      "get": {
        "description": "Download one run's `llms.txt` index as `text/plain`. Any signed-in user, any run.\n\nNot filtered by the caller, on purpose (ARCHITECTURE.md §4.1) — `user_id` is present to\nrequire authentication and is intentionally not passed to the service. `404` if `id`\nnames no run, or if that run has not produced this artifact — see\n`RunService.get_llms_txt` for the two distinguishable `detail`s.\n\n`-> Response`, not `-> str`: this handler's declared contract is that it returns an HTTP\nresponse it built itself, not a value FastAPI should serialize — see `response_class`\nabove and this router's module docstring for why that is not a \"routers are thin\"\nviolation.",
        "operationId": "get_run_llms_txt_runs__id__llms_txt_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "detail": {
                      "type": "string"
                    }
                  },
                  "type": "object"
                }
              }
            },
            "description": "Either there is no run with that id, or that run has not produced this artifact — it is still pending or processing, or it failed. The two cases answer with different `detail` strings; the status is the same because from a client's side both mean there is nothing at this URL."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Run Llms Txt",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/websites/{id}/runs": {
      "get": {
        "description": "List one website's run history, newest first. Any signed-in user, any website.\n\nNot filtered by the caller, on purpose (ARCHITECTURE.md §4.1) — `user_id` is present to\nrequire authentication and is intentionally not passed to the service. `404` if `id`\nnames no website; see `RunService.list_runs`.\n\nThe path parameter is `id`, not `website_id`, because the noun is already in the path\n(ARCHITECTURE.md §10.3) — it shadows the `id` builtin inside this function body, which\nis the accepted cost of the route contract, matching `websites.get_website`.",
        "operationId": "list_runs_websites__id__runs_get",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          },
          {
            "description": "Return only runs with this status. An unrecognized value is a 422.",
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "pending",
                    "processing",
                    "completed",
                    "failed"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Return only runs with this status. An unrecognized value is a 422.",
              "title": "Status"
            }
          },
          {
            "description": "Page size. Values above the maximum (100) are silently clamped to it rather than rejected; `<= 0` is a 422.",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "description": "Page size. Values above the maximum (100) are silently clamped to it rather than rejected; `<= 0` is a 422.",
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "Opaque pagination cursor from a previous page's `next_cursor`. Omit it to fetch the first page. Treat it as a blob — see `Page.next_cursor`.",
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Opaque pagination cursor from a previous page's `next_cursor`. Omit it to fetch the first page. Treat it as a blob — see `Page.next_cursor`.",
              "title": "Cursor"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_RunListItemResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Start a manual crawl of one website. Owner-only — unlike this module's reads.\n\n`202 Accepted`, not `201 Created`: the body describes a `pending` run that has been\nqueued, not one that has finished (see `TriggerRunResponse`). Unlike `list_runs` and\n`get_run`, `user_id` IS passed to the service here, because this is the one write in\nthis router and writes are ownership-checked (ARCHITECTURE.md §4.2).\n\nFour things can stop this short of a `202`: `404` if `id` names no website; `403` if the\ncaller does not own it; `409` if that website already has a run in flight; `429` if the\ncaller is over their concurrent- or daily-run cap. A `503` is also possible, if the run\nwas created but could not be queued — see `RunService.trigger_run` for exactly which\ncheck produces which, and `require_queue_pool` above for the one `503` this router\nitself can raise before the service even runs.",
        "operationId": "trigger_run_websites__id__runs_post",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TriggerRunResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunAlreadyInFlightResponse"
                }
              }
            },
            "description": "This website already has a pending or processing run — of any trigger, manual or scheduled. The response carries that run's id and status."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunLimitExceededResponse"
                }
              }
            },
            "description": "The caller is over their concurrent- or daily-run cap. `detail.scope` says which; `detail.resets_at` is set only for the daily cap."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Run",
        "tags": [
          "runs"
        ]
      }
    },
...
  }
}